Updated to use C++ primitives instead of std library fix with primitives

+++ streets_utils/streets_messages/include/messages/sensor_data_sharing_msg/position_3d.hpp
@@ -9,17 +9,17 @@ namespace streets_utils::messages{
          * @brief LSB = 1/10 micro degree Providing a range of 
          * plus-minus 180 degrees[-1799999999, 1800000001]
          */
-        int64_t _longitude;
+        int _longitude;
         /**
          * @brief LSB = 1/10 micro degree Providing a range of 
          * plus-minus 90 degrees[-900000000, 900000001]
          */
-        int64_t _latitude;
+        int _latitude;
         /**
          * @brief Signed units of 0.1m (10cm), in 2 octets the value 
          * 32767 (0x7FFF) shall indicate an invalid value [-32768,32767]
          */
-        std::optional<int16_t> _elavation;
+        std::optional<int> _elavation;
 
     };
 }
